+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Create a "hello world" plugin without templating (#6, #39)
 - Add status badges to the README (#37)
 - Add a Composer script for PHP linting (#33)
 - Lint the PHP files via GitHub actions (#25, #32)